Trigger .click() event on filter reset link instead of redirecting to its href attribute
The filter action (which includes "?_reset" for reset links) is mapped to a wildcard collection route generated by sfObjectRouteCollection, which requirse a POST method by default.  As such, the URL of the reset link cannot be access by a GET request, be it with jroller or the default admin generator theme.  For additional information, and a possible solution outside the immediate scope of this plugin, see: http://trac.symfony-project.org/ticket/9151